requirements.txt
absl-py==2.1.0
#action-msgs==1.2.1
#action-tutorials-interfaces==0.20.5
#action-tutorials-py==0.20.5
#actionlib-msgs==4.2.4
addict==2.4.0
alabaster==1.0.0
#ament-cmake-test==1.3.11
#ament-copyright==0.12.11
#ament-cppcheck==0.12.11
#ament-cpplint==0.12.11
#ament-flake8==0.12.11
#ament-index-python==1.4.0
#ament-lint==0.12.11
#ament-lint-cmake==0.12.11
#ament-package==0.14.0
#ament-pep257==0.12.11
#ament-uncrustify==0.12.11
#ament-xmllint==0.12.11
angles==1.1.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==25.1.0
azure-core==1.32.0
azure-storage-blob==12.24.1
babel==2.17.0
beautifulsoup4==4.13.3
black==25.1.0
blinker==1.9.0
#builtin-interfaces==1.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.2
#composition-interfaces==1.2.1
ConfigArgParse==1.7
contourpy==1.3.1
cryptography==44.0.0
cv-bridge
cycler==0.12.1
Cython==3.0.12
cython-bbox==0.1.5
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
defusedxml==0.7.1
#demo-nodes-py==0.20.5
#diagnostic-msgs==4.2.4
dill==0.3.9
docutils==0.21.2
#domain-coordinator==0.10.0
einops==0.8.1
#example-interfaces==0.9.3
#examples-rclpy-executors==0.15.3
#examples-rclpy-minimal-action-client==0.15.3
#examples-rclpy-minimal-action-server==0.15.3
#examples-rclpy-minimal-client==0.15.3
#examples-rclpy-minimal-publisher==0.15.3
#examples-rclpy-minimal-service==0.15.3
#examples-rclpy-minimal-subscriber==0.15.3
#exceptiongroup==1.2.2
#executing==2.2.0
faiss-cpu==1.10.0
fastjsonschema==2.21.1
filelock==3.17.0
filterpy==1.4.5
#Flask==3.0.3
flatbuffers==25.1.24
fonttools==4.56.0
fsspec==2025.2.0
fvcore==0.1.5.post20221221
gdown==5.2.0
#geometry-msgs==4.2.4
greenlet==3.1.1
grpcio==1.70.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
#image-geometry==3.2.1
imageio==2.37.0
imagesize==1.4.1
imgaug==0.4.0
#importlib_metadata==8.6.1
#interactive-markers==2.3.2
iopath==0.1.9
ipython==8.32.0
ipywidgets==8.1.5
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lap==0.5.12
#laser-geometry==2.4.0
#launch==1.0.7
#launch-ros==0.19.8
#launch-testing==1.0.7
#launch-testing-ros==0.19.8
#launch-xml==1.0.7
#launch-yaml==1.0.7
lazy_loader==0.4
#lifecycle-msgs==1.2.1
#logging-demo==0.20.5
loguru==0.7.3
#map-msgs==2.1.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
memory-profiler==0.61.0
#message-filters==4.3.5
motmetrics==1.4.0
mpmath==1.3.0
mypy-extensions==1.0.0
narwhals==1.26.0
#nav-msgs==4.2.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
#onnx
#onnx-simplifier
#onnxoptimizer
#onnxruntime
#onnxruntime-gpu
#onnxslim
#open3d==0.18.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
osrf-pycommon
#packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
#pcl-msgs==1.0.0
#pendulum-msgs==0.20.5
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
portalocker==3.1.1
#prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycocotools==2.0.8
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
pyquaternion==0.9.9
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
#python-qt-binding==1.1.2
pytz==2025.1
PyYAML==6.0.2
#qt-dotgraph==2.2.3
#qt-gui==2.2.3
#qt-gui-cpp==2.2.3
#qt-gui-py-common==2.2.3
#quality-of-service-demo-py==0.20.5
#rcl-interfaces==1.2.1
#rclpy==3.3.15
#rcutils==5.1.6
#referencing==0.36.2
#requests==2.32.3
#resource-retriever==3.1.2
#retrying==1.3.4
#rmw-dds-common==1.6.0
#ros2action==0.18.11
#ros2bag==0.15.13
#ros2cli==0.18.11
#ros2component==0.18.11
#ros2doctor==0.18.11
#ros2interface==0.18.11
#ros2launch==0.19.8
#ros2lifecycle==0.18.11
#ros2multicast==0.18.11
#ros2node==0.18.11
#ros2param==0.18.11
#ros2pkg==0.18.11
#ros2run==0.18.11
#ros2service==0.18.11
#ros2topic==0.18.11
#rosbag2-interfaces==0.15.13
#rosbag2-py==0.15.13
#rosgraph-msgs==1.2.1
#rosidl-adapter==3.1.6
#rosidl-cli==3.1.6
#rosidl-cmake==3.1.6
#rosidl-generator-c==3.1.6
#rosidl-generator-cpp==3.1.6
#rosidl-generator-py==0.14.4
#rosidl-parser==3.1.6
#rosidl-runtime-py==0.9.3
#rosidl-typesupport-c==2.0.2
#rosidl-typesupport-cpp==2.0.2
#rosidl-typesupport-fastrtps-c==2.2.2
#rosidl-typesupport-fastrtps-cpp==2.2.2
#rosidl-typesupport-introspection-c==3.1.6
#rosidl-typesupport-introspection-cpp==3.1.6
#rpds-py==0.22.3
#rpyutils==0.2.1
#rqt-action==2.0.1
#rqt-bag==1.1.5
#rqt-bag-plugins==1.1.5
#rqt-console==2.0.3
#rqt-graph==1.3.1
#rqt-gui==1.1.7
#rqt-gui-py==1.1.7
#rqt-msg==1.2.0
#rqt-plot==1.1.2
#rqt-publisher==1.5.0
#rqt-py-common==1.1.7
#rqt-py-console==1.0.2
#rqt-reconfigure==1.1.2
#rqt-service-caller==1.0.5
#rqt-shell==1.0.2
#rqt-srv==1.0.3
#rqt-topic==1.5.0
#rtsp-camera-publisher==0.0.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
#sensor-msgs==4.2.4
#sensor-msgs-py==4.2.4
#shape-msgs==4.2.4
shapely==2.0.7
six==1.17.0
snakeviz==2.2.2
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.38
#sros2==0.10.5
#stack-data==0.6.3
#statistics-msgs==1.2.1
#std-msgs==4.2.4
#std-srvs==4.2.4
#stereo-msgs==4.2.4
supervision @ git+https://github.com/roboflow/supervision.git@39a12db1d4df67bfc066ef860c7991888bfaaa0f
sympy==1.13.3
tabulate==0.9.0
#teleop-twist-keyboard==2.4.0
#tensorboard
#tensorboard-data-server
#tensorrt
#tensorrt-cu12
#tensorrt-cu12-bindings
#tensorrt-cu12-libs
#termcolor==2.5.0
#tf2-geometry-msgs==0.25.9
#tf2-kdl==0.25.9
#tf2-msgs==0.25.9
#tf2-py==0.25.9
#tf2-ros-py==0.25.9
#tf2-tools==0.25.9
#threadpoolctl==3.5.0
#tifffile==2025.1.10
#tomli==2.2.1
#topic-monitor==0.20.5
torch==2.4.1
torchreid==0.2.5
torchvision==0.19.1
tornado==6.4.2
tqdm==4.66.5
#traitlets==5.14.3
#trajectory-msgs==4.2.4
#triton==3.0.0
#turtlesim==1.4.2
#typing_extensions==4.12.2
#tzdata==2025.1
ultralytics==8.3.24
ultralytics-thop==2.0.10
#unique-identifier-msgs==2.2.1
#urllib3==2.3.0
#vidgear==0.3.3
#visualization-msgs==4.2.4
#wcwidth==0.2.13
#Werkzeug==3.0.6
#widgetsnbextension==4.0.13
#win32_setctime==1.2.0
xmltodict==0.14.2
#yacs==0.1.8
#zipp==3.21.0
